Do not automagically online memory on s390x (bsc#1127557).
On s390 memory that is not online is typically 'reserve' memory that
should be only activated by the system administrator.

There is no way to detect reserve memory from information provided by
the kernel so the state of the art is to not activate memory
automagically if the platform is s390.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>